fix: locks Typescript dependency version
Typescript version 4.8 and above are incompatible with the `@types/mongodb`
package. This change locks the Typescript dependency to v4.7.4 the last
version that is compatible as the old dependency was "^4.1.2" it allowed
incompatible versions to be installed.

The proper fix would be to move to mongodb >= 4.0.0 which bundles its
own Typescript types and drop `@types/mongodb` then unlock the Typescript
dependency.