feat: force-close long-lived connections that prevent the server from shutting down
Some browsers .. ** cough, cough -- CHROME ** .. hold open phony
"in-flight" requests, presumably for performance reasons. They'll send
headers and request body once the client decides it has a request to
actually send, but to the server, this phony open request looks like a
legitimate open connection that has begun processing. So in development,
err on the side of forcing connections to terminate early. In ECS
environments, We should be insulated from this behavior by our load
balancer.